requirements.txt
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.15.0
numpy>=1.21.0
wandb>=0.12.0
Pillow>=8.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
opencv-python>=4.5.0
timm>=0.5.0
tensorboard>=2.7.0
pycocoevalcap
nltk>=3.6.0 